Pull out all the allowance-related code into separate module
