+++ contracts/cw20-base/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod allowances;
 pub mod contract;
 pub mod msg;
 pub mod state;
